Add all content markdown files into the purge
This results in a ~3k increase of built JS (0.6k gzipped) because
additonal classes match. That's not great, but also no big deal after
all, and it allows us to use font awesome icons and maybe very special
stylings directly inside the content without being careful, so this,
overall, is a win.